requirements.txt
python-dotenv
assemblyai
openai
anthropic
gradio

